+++ thoth-api/src/model/mod.rs
@@ -22,9 +22,12 @@ pub const ROR_DOMAIN: &str = "https://ror.org/";
 #[serde(rename_all = "SCREAMING_SNAKE_CASE")]
 #[strum(serialize_all = "lowercase")]
 pub enum LengthUnit {
+    #[cfg_attr(feature = "backend", graphql(description = "Millimetres"))]
     #[default]
     Mm,
+    #[cfg_attr(feature = "backend", graphql(description = "Centimetres"))]
     Cm,
+    #[cfg_attr(feature = "backend", graphql(description = "Inches"))]
     In,
 }
 
@@ -37,8 +40,10 @@ pub enum LengthUnit {
 #[serde(rename_all = "SCREAMING_SNAKE_CASE")]
 #[strum(serialize_all = "lowercase")]
 pub enum WeightUnit {
+    #[cfg_attr(feature = "backend", graphql(description = "Grams"))]
     #[default]
     G,
+    #[cfg_attr(feature = "backend", graphql(description = "Ounces"))]
     Oz,
 }
 
@@ -394,7 +399,7 @@ macro_rules! crud_methods {
         fn from_id(db: &$crate::db::PgPool, entity_id: &Uuid) -> ThothResult<Self> {
             use diesel::{QueryDsl, RunQueryDsl};
 
-            let mut connection = db.get().unwrap();
+            let mut connection = db.get()?;
             match $entity_dsl
                 .find(entity_id)
                 .get_result::<Self>(&mut connection)
@@ -405,7 +410,7 @@ macro_rules! crud_methods {
         }
 
         fn create(db: &$crate::db::PgPool, data: &Self::NewEntity) -> ThothResult<Self> {
-            let mut connection = db.get().unwrap();
+            let mut connection = db.get()?;
             match diesel::insert_into($table_dsl)
                 .values(data)
                 .get_result::<Self>(&mut connection)
@@ -425,7 +430,7 @@ macro_rules! crud_methods {
         ) -> ThothResult<Self> {
             use diesel::{Connection, QueryDsl, RunQueryDsl};
 
-            let mut connection = db.get().unwrap();
+            let mut connection = db.get()?;
             connection.transaction(|connection| {
                 match diesel::update($entity_dsl.find(&self.pk()))
                     .set(data)
@@ -443,7 +448,7 @@ macro_rules! crud_methods {
         fn delete(self, db: &$crate::db::PgPool) -> ThothResult<Self> {
             use diesel::{QueryDsl, RunQueryDsl};
 
-            let mut connection = db.get().unwrap();
+            let mut connection = db.get()?;
             match diesel::delete($entity_dsl.find(&self.pk())).execute(&mut connection) {
                 Ok(_) => Ok(self),
                 Err(e) => Err(ThothError::from(e)),