+++ thoth-errors/src/lib.rs
@@ -85,15 +85,17 @@ pub enum ThothError {
         "Price values must be greater than zero. To indicate an unpriced Publication, omit all Prices."
     )]
     PriceZeroError,
+    #[error("Publication Date is required for Active, Withdrawn, and Superseded Works.")]
+    PublicationDateError,
     #[error("{0}")]
     RequestError(String),
     #[error("{0}")]
     GraphqlError(String),
     #[error("Withdrawn Date must be later than Publication Date.")]
     WithdrawnDateBeforePublicationDateError,
-    #[error("Withdrawn Date can only be added to an Out of Print or Withdrawn From Sale Work.")]
+    #[error("Withdrawn Date can only be added to a Superseded or Withdrawn Work.")]
     WithdrawnDateError,
-    #[error("An Out of Print or Withdrawn From Sale Work must have a Withdrawn Date.")]
+    #[error("A Superseded or Withdrawn Work must have a Withdrawn Date.")]
     NoWithdrawnDateError,
 }